Track active conversations alongside online status
How:
Instead of simply tracking online users via their ids in a list, track
their active conversations as well. Thus keep a list of objects
which stores a user's id as well as their active conversation.

Why:
All users should be aware of other users' active conversations so that
messages sent to/ from a particular user can be marked as unread/read
depending on whether or not the sender and reciever are in an active
conversation with one another.