Read messages via patch rather than post request
Messages can be read in one go simply with the conversation id rather
than updating each message in the db individually.
Updated messages do not need to be returned by the api call, rather the
store can be updated simply after confirmation that messages for a
particular conversation were updated in the db.