Account for message recipient's current search
When sending a message to a new conversation, account for the fact the
other user maybe currently searching for you. If this is the case, then
sending a new message will create a new conversation alongside the
"fake" conversation that will pop-up on their sidebar whilst they are
searching. This is not the desired behaviour as we do not want duplicate
conversations. This commit overwrites the existing "fake" conversation
in this edge case scenario.