Fix bug with avatar disappearing in active convos
Last commit made it so avater would disappear when the other user in the
conversation would send a message. This was because each message being
sent was replacing the last read message and not accounting for who was
sending the message and who was receiving it. With this change, the last
message read would only be updated for the user sending the message if
in an active conversation.