Make junction model between message and user
Due to the possibility of a conversation having more than two users, it
is no longer sufficient to just have a "read" attribute in the message
model. Each user must have their own "read" attribute for each message
that is sent in any conversation they are a part of. This can be
acheived with a junction model between user and message.